fix: correctly convert to openpgp ecdsa key representation
